#7: Move Square, Colour, Piece to enums.rs
