+++ src/main.rs
@@ -1,6 +1,7 @@
 #![allow(dead_code)]
 
 mod aliases;
+mod enums;
 mod magic;
 mod masks;
 mod positions;
